Função adicionada
Foi adicionado um tipo de sistema de pontuação, onde será possível para o jogador ver sua pontual na tela de seleção de categoria e de jogo.
adicionado um novo script ScoreManager;
adicionado game Object com nome de ScoreManager nas telas Menu selecao e tela de jogo;
Testes feitos, erros removidos palavras alteradas;

+++ Word Game/Assets/Scripts/Cobjetos/BoardData.cs
@@ -45,7 +45,7 @@ public class BoardData: ScriptableObject
         }
     }
     
-    public float timeSeconds;
+    public int timeSeconds;
     public int Columns = 0;
     public int Rows = 0;
